+++ .changeset/warm-steaks-approve.md
@@ -1,0 +1,5 @@
+---
+"inngest": minor
+---
+
+Added h3 framework server handler
